pyproject.toml
[project]
name = "item-archive-publisher"
version = "0.1.0"
description = "A simple Python package"
requires-python = ">=3.11"
dependencies = [
    "pycryptodome>=3.23.0",
    "pyyaml>=6.0.2",
    "rich>=14.1.0",
]
